Add command line parser
Using boost program_option to interpret the command line arguments.

The idea is to create an argument object filled with default values.
Then the configuration file is parsed and the contained values updated.
At the end, command line arguments are parsed and the arg object will be
updated accordingly.

part of #4